Move logic out of app.js
