+++ frontend/smart-kickers-game/src/components/Game/GameStatistics/GameStatistics.js
@@ -8,12 +8,10 @@ import FinalScores from './StatisticsItems/FinalScores.js';
 import FastestShot from './StatisticsItems/FastestShot.js';
 import ManualChangedGoals from './StatisticsItems/ManualChangedGoals.js';
 import TeamIcons from './StatisticsItems/TeamIcons.js';
-import { useNavigate } from 'react-router-dom';
 import { useGameDataContext } from '../../../contexts/GameDataContext';
 
 function GameStatistics() {
   const [statistics, setStatistics] = useState(null);
-  const navigate = useNavigate();
   const finalScores = useGameDataContext().finalScores;
 
   const handleGetStatistics = async () => {
